Implement foundational structure for LTX video processing: Add Dockerfile and Makefile for project setup, including repository cloning and dependency installation. Introduce new model components such as attention mechanisms, autoencoders, and various embedding classes to enhance video generation capabilities. Refactor existing modules for improved maintainability and flexibility, ensuring compatibility with Q8 optimizations and dynamic processing strategies.
