+++ README.md
@@ -1,2 +1,11 @@
 # inferno
-Not Diffusers 
+
+Not Diffusers
+
+## Installation
+
+To set up the project, simply run:
+
+```bash
+make setup
+```
