--- inferno/ltx_engine/model/__init__.py
@@ -1,1 +1,0 @@
-from . import autoencoder, patchifier, transformer, utils
